requirements.txt
cachetools==5.3.0
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
isodate==0.6.1
moviepy==1.0.3
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
pillow==9.4.0
proglog==0.1.10
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==12.1.3
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15